admin: datafeed renaming
